requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pydantic>=2.0.0
langgraph>=0.0.1
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
